Version 31
7 Translated languages (Still have Russian, Korean, and  Thai, and
Norwegian to go)



This version is a rewrite of the GUI.

The code is much tighter and better structured.



The following configuration option are added to the client side.

Vertical location of the

* prepared spell display.

* spell casting bar

* mana display

The items you can configure on the client side start with "gui".

The Casting Bar.

The casting bar is now graphical and also has a custom icon in the
middle for Glorious recall which shows if you are teleporting to a bed
or worldspawn.



Prepared spell area:
This now shows the spells you will prepare if you look up or down above
and below the current prepared spell.

The lower prepared spell also fades out instead of blinking off.



More mana regeneration options were added on the server side
configuration.

The operator can set armor to have a small regeneration that can take
them to 100% mana that takes about 50 game days in leather redstone
armor and 25 days in redstone suit armor.

The operator can set a small regeneration when the player sleeps in a
bed that can take them all the way to 100% mana.  The default takes
about 30 game days.



Flying speed now depends on the armor worn.   No redstone armor is the
slowest.  There is a speed bonus per piece of redstone leather or
redstone suit armor.

Flying spee